pyproject.toml
[tool.poetry]
name = "arista"
version = "0.1.0"
description = "Arista - crypto signals"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pydantic = "^2.7.1"
psycopg2-binary = "2.9.5"
sqlmodel = "^0.0.21"
alembic = "^1.13.2"
jupyterlab = "^4.2.4"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
isort = "^5.13.2"
ruff = "^0.4.9"

black = "^24.10.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-exclude = ["arista/alembic/*"]

[tool.isort]
skip_glob = ["arista/alembic/*"]

[tool.poetry.scripts]
sync_coinglass = "arista.scripts.coinglass:main"
sync_cmc = "arista.scripts.cmc:main"


